backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-docx
pypdf
google-auth
requests
python-dotenv
